Added credit meter on chat page

+++ backend/app/models/module_assignment.py
@@ -9,7 +9,7 @@ class ModuleAssignment(db.Model):
     # Assuming userID is a foreign key to User.userID
     userID = db.Column(db.Integer, db.ForeignKey('dbo.Users.userID'), nullable=False)
     moduleID = db.Column(db.String(50), nullable=False)
-    studentCredits = db.Column(db.Integer, nullable=True)
+    studentCredits = db.Column(db.Float, nullable=True)
 
     # Define the relationship to User
     user = db.relationship('User', backref=db.backref('module_assignments', lazy=True))
